+++ Tests/SentryTests/Networking/SentryNSURLRequestBuilderTests.swift
@@ -66,7 +66,9 @@ class SentryNSURLRequestBuilderTests: XCTestCase {
     private func givenEnvelopeWithInvalidData() -> SentryEnvelope {
         let sdkInfoWithInvalidJSON = SentrySdkInfo(
             name: SentryInvalidJSONString() as String,
-            andVersion: "8.0.0"
+            version: "8.0.0",
+            integrations: [],
+            features: []
         )
         let headerWithInvalidJSON = SentryEnvelopeHeader(
             id: nil,