+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - SwiftUI time for initial display and time for full display (#4596)
 
+### Improvements
+
+- Improve compiler error message for missing Swift declarations due to APPLICATION_EXTENSION_API_ONLY (#4603)
+
 ## 8.42.0-beta.2
 
 ### Fixes