+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 
 - `SentrySdkInfo.packages` should be an array (#4626)
+- Use the same SdkInfo for envelope header and event 
 
 ### Internal