+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Improvements
+
+- Improve compiler error message for missing Swift declarations due to APPLICATION_EXTENSION_API_ONLY (#4603)
+
+## 8.42.0-beta.2
+
 ### Fixes
 
 - Fix GraphQL context for HTTP client error tracking (#4567)